requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT==2.8.0
pytz
sqlparse
psycopg2-binary
python-dotenv
requests
django-allauth
jwt
django-crispy-forms
gunicorn